[JRO] Setup teaspoon
Because of how much javascript there is going on in here, I reckon, wtf,
might as well TDD most of this instead of cowboy code this into a
heaping mess of shit. I've had enough of dumpster-fire codebases so ...
might as well get a proper JS test harness hooked up.

This is using https://github.com/modeset/teaspoon and while it's a
little bit of a pain in the ass to get working, it's pretty nice so far.